backend/requirements.txt
faster-whisper>=1.0.0
yt-dlp>=2024.1.0
google-api-python-client>=2.0.0
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
pgvector>=0.2.0
httpx>=0.27.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
